+++ src/main/java/storage/InternalPath.java
@@ -1,3 +1,5 @@
+package storage;
+
 import java.io.File;
 import java.nio.file.Path;
 import java.util.Arrays;
