+++ src/main/java/storage/InternalStorage.java
@@ -1,3 +1,5 @@
+package storage;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
