+++ src/main/java/util/events/EventEmitter.java
@@ -1,8 +1,10 @@
+package util.events;
+
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * A class capable of emitting events to event listeners.
+ * A class capable of emitting utils.events to event listeners.
  */
 public class EventEmitter<T> {
     private List<EventListener<? super T>> eventListeners = new ArrayList<>();
@@ -18,7 +20,7 @@ public class EventEmitter<T> {
     }
 
     /**
-     * Adds an event listener to listen to events by this instance.
+     * Adds an event listener to listen to utils.events by this instance.
      * @param listener The listener to add.
      */
     public void addEventListener(EventListener<? super T> listener) {
@@ -26,7 +28,7 @@ public class EventEmitter<T> {
     }
 
     /**
-     * Removes an event listener to remove it from listening to events by this instance.
+     * Removes an event listener to remove it from listening to utils.events by this instance.
      * @param listener The listener to remove.
      */
     public void removeEventListener(EventListener<? super T> listener) {
